refactor(telemetry): move from pkg/util/telemetry to internal/telemetry
Package `pkg/util/telemetry` is not intended to expose; therefore, this PR moves it to
`internal/telemetry`.
