+++ internal/metarepos/telemetry.go
@@ -5,8 +5,8 @@ import (
 
 	"go.opentelemetry.io/otel/metric"
 
+	"github.com/kakao/varlog/internal/telemetry"
 	"github.com/kakao/varlog/pkg/types"
-	"github.com/kakao/varlog/pkg/util/telemetry"
 )
 
 type telemetryStub struct {
